reduxTypes replaced file actionTypes wit enum ActionTypes.
