--- src/redux/DataStructure/actionTypes.ts
@@ -1,1 +1,0 @@
-export const SET_DATA_STRUCTURE = 'SET_DATA_STRUCTURE';
